Implemented a 'feature' navigation record in a separate Gradle module
This showcase the ability to scale an app by feature modules, and those modules defining the nav records and UI and exposing the back stack key to the consumer so the UI to navigate to it is possible.

It also demonstrates the nav records and their UI can have their scoped dependencies and that those dependencies can use the host activity.